Code/projects/RadarLead/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
streamlit==1.28.0
streamlit-folium==0.15.1
folium==0.15.1
pandas==2.1.3
requests==2.31.0
playwright==1.40.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
# AI enrichment
anthropic>=0.40.0
# SaaS billing
stripe>=7.0.0

